% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tradeoff}
\alias{plot_tradeoff}
\title{Plot DALYs per 100,000 by infection type}
\usage{
plot_tradeoff(dataset = germany_burden, country = "Germany")
}
\arguments{
\item{dataset}{Data frame of hai_type and dalys_per100k (germany_burden or eu_eea_burden)}

\item{country}{Title label ("Germany" or "EU/EEA")}
}
\description{
Plot DALYs per 100,000 by infection type
}
