% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_burden.R
\name{table_burden}
\alias{table_burden}
\title{Present burden table (Germany by default)}
\usage{
table_burden(dataset = germany_burden)
}
\arguments{
\item{dataset}{Data frame to display (default: germany_burden)}
}
\description{
Present burden table (Germany by default)
}
