% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R, R/dataset.R
\docType{data}
\name{eu_eea_burden}
\alias{eu_eea_burden}
\title{EU/EEA burden (DALYs per 100,000) by infection type}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 5 columns.

Same columns as \code{germany_burden}; values are per 100,000.
}
\source{
Zacher et al., Euro Surveill 2019.
}
\usage{
eu_eea_burden

eu_eea_burden
}
\description{
EU/EEA burden (DALYs per 100,000) by infection type

EU/EEA HAI burden (per 100,000 population, medians)
}
\keyword{datasets}
