% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{metadata_bhai}
\alias{metadata_bhai}
\title{Metadata and key assumptions (provenance, method, McCabe proxy)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
metadata_bhai
}
\description{
Metadata and key assumptions (provenance, method, McCabe proxy)
}
\keyword{datasets}
