% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tradeoff}
\alias{plot_tradeoff}
\title{Plot DALYs per 100,000 by infection type}
\usage{
plot_tradeoff(data, region = "Region")
}
\arguments{
\item{data}{Data frame with columns `hai_type` and `dalys_per100k`.}

\item{region}{Character string specifying region name (e.g. "Germany" or "EU/EEA").}
}
\value{
A ggplot object showing DALYs by infection type.
}
\description{
Plot DALYs per 100,000 by infection type
}
\examples{
plot_tradeoff(asg4::germany_burden, "Germany")
}
