% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_burden.R
\name{table_burden}
\alias{table_burden}
\title{Build a formatted burden table}
\usage{
table_burden(data)
}
\arguments{
\item{data}{Tibble with hai_type, cases, deaths, dalys, dalys_per100k.}
}
\value{
A gt table.
}
\description{
Build a formatted burden table
}
